+++ extensions/raycast-wallpaper/src/hooks/hooks.ts
@@ -6,8 +6,64 @@ import { cachePicture, checkCache } from "../utils/common-utils";
 import axios from "axios";
 import Style = Toast.Style;
 
+const backupWallpapers: RaycastWallpaper[] = [
+  {
+    title: "Autumnal Peach",
+    url: "https://www.raycast.com/uploads/wallpapers/autumnal-peach.png",
+  },
+  {
+    title: "Blossom",
+    url: "https://www.raycast.com/uploads/wallpapers/blossom-2.png",
+  },
+  {
+    title: "Blushing Fire",
+    url: "https://www.raycast.com/uploads/wallpapers/blushing-fire.png",
+  },
+  {
+    title: "Bright Rain",
+    url: "https://www.raycast.com/uploads/wallpapers/bright-rain.png",
+  },
+  {
+    title: "Floss",
+    url: "https://www.raycast.com/uploads/wallpapers/floss.png",
+  },
+  {
+    title: "Glass Rainbow",
+    url: "https://www.raycast.com/uploads/wallpapers/glass-rainbow.png",
+  },
+  {
+    title: "Good Vibes",
+    url: "https://www.raycast.com/uploads/wallpapers/good-vibes.png",
+  },
+  {
+    title: "Moonrise",
+    url: "https://www.raycast.com/uploads/wallpapers/moonrise.png",
+  },
+  {
+    title: "Ray of Lights",
+    url: "https://www.raycast.com/uploads/wallpapers/ray-of-lights.png",
+  },
+  {
+    title: "Rose Thorn",
+    url: "https://www.raycast.com/uploads/wallpapers/rose-thorn.png",
+  },
+];
+
 export const getRaycastWallpaperList = () => {
   const [raycastWallpapers, setRaycastWallpapers] = useState<RaycastWallpaper[]>([]);
+
+  const storeRaycastWallpapers = (wallpapers: RaycastWallpaper[]) => {
+    setRaycastWallpapers(wallpapers);
+
+    //cache list
+    LocalStorage.setItem(LocalStorageKey.WALLPAPER_LIST_CACHE, JSON.stringify(wallpapers));
+
+    wallpapers.forEach((value) => {
+      if (!checkCache(value)) {
+        cachePicture(value);
+      }
+    });
+  };
 
   const fetchData = useCallback(async () => {
     //get wallpaper list
@@ -26,20 +82,11 @@ export const getRaycastWallpaperList = () => {
         },
       })
         .then((axiosRes) => {
-          const _raycastWallpaper = axiosRes.data as RaycastWallpaper[];
-          setRaycastWallpapers(_raycastWallpaper);
-
-          //cache list
-          LocalStorage.setItem(LocalStorageKey.WALLPAPER_LIST_CACHE, JSON.stringify(_raycastWallpaper));
-
-          _raycastWallpaper.forEach((value) => {
-            if (!checkCache(value)) {
-              cachePicture(value);
-            }
-          });
+          storeRaycastWallpapers(axiosRes.data as RaycastWallpaper[]);
         })
-        .catch((error) => {
-          showToast(Style.Failure, String(error));
+        .catch(() => {
+          // In the case of an error, use the backup list.
+          storeRaycastWallpapers(backupWallpapers);
         });
     } catch (e) {
       await showToast(Style.Failure, String(e));