+++ extensions/raycast-wallpaper/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Raycast Wallpaper Changelog
 
+## [Add backup] - 2022-11-08
+
+- Add a hardcoded list of wallapapers in case there's an error fetching the wallpapers from the API.
+
 ## [Update grid layout] - 2022-10-15
 
 - Update grid layout